requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pycodestyle==2.10.0
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
